pyproject.toml
[tool.poetry]
name = "letter-of-rec"
version = "0.1.0"
description = "Tools for managing and redacting student information in letters of recommendation"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "lor"}]

[tool.poetry.dependencies]
python = "^3.12"
python-docx = "^1.2.0"
mammoth = "^1.11.0"
openai = "^2.8.1"
python-dotenv = "^1.2.1"
click = "^8.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"

[tool.poetry.scripts]
lor = "lor.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
